refactor: trim thread header tests and helpers
Drop the unrelated AGENTS.md note and cloudProjectAliases move, collapse
the header/prefs/backend tests to the real failure modes, inline the
one-caller ThreadContextMenuPopup and ThreadProjectIndicator helpers.
